requirements.txt
streamlit
numpy
pillow
tensorflow
chromadb
python-dotenv
google-generativeai
fastapi
uvicorn
python-multipart
